Added option to prevent modular to take control of the compiler's JS output - instead it only emits arguments for haxe-split. Added helpers for Kha to hook in Kha build process.
